cdas_backend/requirements.txt
asgiref>=3.7,<4.0
Django>=5.0,<5.1
djangorestframework>=3.16,<3.18
sqlparse>=0.5,<0.6
python-dotenv>=1.0,<2.0

# Auth JWT
djangorestframework-simplejwt>=5.3,<6.0

# Auth OAuth
django-allauth>=65.0,<66.0
dj-rest-auth>=5.0,<5.1

# Images
Pillow>=10.0,<11.0
django-imagekit>=5.0,<6.0

# Documents
PyPDF2>=3.0,<4.0
python-docx>=0.8,<0.9
openpyxl>=3.0.10,<4.0

#Celery dependances
celery>=5.3,<5.4
django-celery-beat>=2.6,<2.7
django-celery-results>=2.6,<2.7

# Base de donnees
psycopg2-binary>=2.9,<3.0

#Email
# django-ses>=3.5,<4.0
# sendgrid>=6.10,<7.0
# django-anymail>=10.2,<11.0

#Requests
requests>=2.31,<3.0
httpx>=0.26,<1.0
urllib3>=2.0,<3.0

#Documentation APIs
drf-spectacular>=0.27,<1.0

#Securite
cryptography>=42.0,<43.0
argon2-cffi>=23.1,<24.0